We've added more details about security update in release summary
Notes: This version also includes security fixes so we recommend to update and use the latest version.  A security issue allowing registered users to escalate their privileges to a forum administrator was fixed. Look through the forum members and filter by administrator group, look for any suspicious administrator-level accounts that could be a backdoor and remove them accordingly.
https://wpforo.com/community/wpforo-announcements/wpforo-1-5-2-is-released/